Script Reversion
Went through the changes from Wertercatt with the knowledge that the program they used was probably shitting itself over the formatting stuff in these, un-Americanizing what got the Yankee treatment from it, and undoing proper grammar in dialog where it made sense for the character. Yes, the narrator is a character too.

Didn't fine-tooth comb the Thes because I know someone on the main branch already touched that up.

Also, some little fixes of my own sprinkled in.