localization hotfix
So there's a bug with localization caching that we've run into because of the typo originally in the file's name. A blank copy of the old name needs to be included so the cache from it doesn't override the new file.

Reported this bug for GMod but who knows when/if it'd get fixed, and the localization was already put up on Steam so this would affect *everyone* who has downloaded that update and subsequently played Jazztronauts (or possibly just started up GMod, unclear on when they're cached).