selector/fgd update
Updated so selector turns towards player regardless of angle set on the entity (the old code required the selector be rotated to 0 270 0 to work properly)

Added "facing" property for the selector. Intended to control the axis (if any) that the rotation happens on, but right now it's either rotate as it is now or disable.

More fgd fixups: removed a lot of useless origins from stuff, and added the different options for specific facts to the fact screens, rather than having to type them out yourself.