Player Dialog consistency fix
Fixes inconsistency in dialog box model used between player dialog that's hard-scripted vs. player dialog picked from options.

Current behavior is that the player proxy is used if it is valid and the player doesn't currently have PAC enabled, otherwise the player themself is used.

If we can get the player proxy to play nicely with PAC, that part of the conditional can get removed, but right now there's too many possible break points with PAC.