Disable Auto-Switching off Jazz weapons
Prevents picking up a crossbow, .357, etc. from yoinking a player out of the Jazz weapons. Most for functional reasons (don't want to interrupt the player snatching if they happen to pick up a weapon, or if they're trying to call the bus, or bore with Stan, or not die to fall damage from Run, etc.)