Tweak hub transition timing + network fixes
This changes it so the transition when spawning in the hub is consistent. To do this, I made OnClientInitialized call clientside as well, to know when fully loaded into the world.

I also added an OnClientInitialized block for the singleplayer check, since I noticed it similarly works better this way. There's likely other things to find later.