Remove player_collide, check for stuck players
Liked the idea, but with more testing, seems to be a lot more effort and fuckiness than expected for a rather niche convar. Collision enabling is now just tied to PvP directly.

Also, when enabling PvP, all players get an unstuck check, in case they happened to be inside each other