added checks for aspect ratios (16:9, 16:10 & 4:3 so far))
thanks RDBomber for setting the foundation. wrote slightly cleaner code but commented out a check for resolutions lower than 720p because it broke everything for some reason.

the 720p or lower check outputs a LUA error whereas removing it no longer outputs an error and idfk why, it *looks* like it shouldn't break anything, but it does, and that's why i commented it out.

Co-Authored-By: RDBomber (OtherElement) <62676175+rdbomber@users.noreply.github.com>