Remove underscore check in strict regex
Checked myself, it was (un)matching 95 items. Unfortunately, 9 of those (being generous) are false positives which will now be excluded. But excluding the junk seems worth it.

Code could be simplified to remove the regex string, but we may want to make it a regex again later, and there's some added complexity in staying case-insensitive.